% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelprobit.R
\name{probit_model}
\alias{probit_model}
\title{Modelo Probit}
\usage{
probit_model(formula, data)
}
\arguments{
\item{formula}{(vector) un vector que contiene las variables prdictoras en relacion de la variable respuesta}

\item{data}{(data.frame) un data frame que contiene los datos con los que se trabajará.}
}
\description{
Analisis de regresion con variables de respuesta binaria.
}
\examples{
\dontrun{
# Cargar los datos
datos <- read.csv("/Users/marif/Downloads/dataset1.csv")

# Eliminar filas con valores faltantes
datos <- na.omit(datos)
# Variables
y <- datos$presencia
x <- datos[, c("rough", "julmint", "pptcv", "evi")]

# Ajustar modelo Probit
model <- probit_model(y ~ rough + julmint + pptcv + evi, data = data.frame(y, x))
# Obtener predicciones
predictions <- model$predict(data.frame(x))
# Imprimir coeficientes estimados y predicciones
print(model$coefficients)
print(predictions)
}
}
